requirements.txt
h2o>=3.46.0
evidently==0.5.0
pandas>=1.5.0
scikit-learn>=1.3.0
numpy>=1.24.0
mlflow
requests
python-dotenv